+++ lib/src/listener.dart
@@ -1,0 +1,8 @@
+class Listener {
+  String path;
+  Function callback;
+  RegExp regExp;
+  Map<dynamic, dynamic> locals;
+
+  Listener(this.path, this.callback, this.regExp, this.locals);
+}